requirements.txt
certifi==2020.12.5
cycler==0.10.0
Cython==0.29.22
kiwisolver==1.3.1
matplotlib==3.4.1
numpy==1.20.2
Pillow>=8.1.2
pycocotools @ git+https://github.com/gautamchitnis/cocoapi.git@20291f19c46a8d11935862bc9e449a1b72ec25ed#subdirectory=PythonAPI
pyparsing==2.4.7
python-dateutil==2.8.1
six==1.15.0
torch==1.8.1
torchvision==0.9.1
typing-extensions==3.7.4.3
wincertstore==0.2
tqdm~=4.60.0
torchsummary~=1.5.1